GCC -ansi enables long long and 64-bit support, so allows it
Even if not a Standard C89 feature, GCC when invoked in -ansi mode
honours the long long construction; and this is used in some code,
particularly to get access to 64-bit arithmetics while at the same time
disabling most of the annoying GCC extensions and get better diagnostics.

The declaration of int64_t in <stdint.h> depends on __LONG_LONG_SUPPORTED.
Remove its dependency on __STRICT_ANSI, the macro associated to -ansi.

See details in the thread
http://groups.google.com/group/minix3/browse_thread/thread/5cff67854ba924b1