fix advent for objdir feature of BSD make
This is a sequel of bug #486 (rev. 7445, 22 Jun).
This bugfix did not solve all the problems, particularly with respect to the
advtext.h file which is created in ${.OBJDIR}, not in ${.CURDIR}.
Also, a further modification to avoid warnings was not compatible with the
objdir feature either. Both issues fixed.