Fix ash to be cross-compiled easily
ash needs several helper programs (mk*) to be run while compiling; when
cross-compiling, this means using the host C compiler, rather than the
normal cross-compiler targetting MINIX. Also, this impacts the
generation of the yacc-generated header.

A special point about cross-compiling is worth noting: since the
various mk* programs run on the host platform, they use the host include
files. In the case of mksignames, this means the built ash behaves with
respect of the host view of signals, which may, or may not, matches the
one of the target. This particular problem is long-known for shells
(well described in http://www.mail-archive.com/bug-bash@gnu.org/msg00831.html).
There are numerous ways to fix that, like using a MINIX-specific version
of the signals rather than the standard C' <signal.h>; see the handling
done for the keyboard maps in drivers/tty/keymaps for an example.