Fix <bsd.lib.mk> to be cross-compiled easily
The <bsd/minix.*.mk> files were lifted from NetBSD back in 2009;
Then, minix.lib.mk was heavily washed out to adapt itself to the
MINIX tree (the present file is about 15% of the size of the
original!), that is, without the shared libraries stuff.

However, doing so, a key part of the cross-compiling infrastructure
was dropped, and to be precise the hability when DESTDIR is defined
to use a different place for the include files; as a result the
behaviour of the cross-compiler is different when it compiles PROGrams
(it searches the include files inside $(DESTDIR)/usr/include, which
is where $(MAKE) includes placed them) and when it compiles LIBraries
(it searches the includes files the way it is configured by default).
This fix aligns the behaviours (to the one of bsd.prog.mk)

Also two variables useful in cross-compilation cases (_GCC_CRTDIR and
_GCC_LIBGCCDIR) are initialized to some default value.