Fix readfs to make it working with V3 file systems
Readfs(1) is a MINIX-specific tool which is around for a long time.
However, while it has been fixed to allow compilation, it was never made to
work with real V2. much less V3, filesystems. In the current form, it just
allows to deal with those baroque "pseudo-V1" file systems which
mkfs(1) -1 can create.

This patch allows it to work with real V3 file systems.