New susp'd condition FP_BLOCKED_ON_BGIO
Indicate that a system call is suspended because the process
is in a background process group, and it tried to do I/O on
its controlling terminal.
Unlike normal I/O on a tty, nothing is stored inside the tty
driver; at unblocking time, the whole system call will be
restarted from the starting point in get_work().

Also create a new FSTATE, 'B', in procfs.