Support for TOSTOP
TOSTOP prevents background processes to write on their
controlling terminal. The actual action is the same as with
READ or IOCTL. But since it happens within VFS, we must remember
the current value of that TOSTOP flag (which is part of the
local flags of the termios structure); so we also intercept
the IOCTL TCSETS which is the only way to affect this flag.
Note the flag is actually remembered by the session leader.