Update_bootcfg: Updated
Manual page.
Added several options.
Ability to remove a broken minix_latest,
and to fix it to point to the real latest still on disk.
Dropped INHERIT stuff, following idea from David.
Sorted out the mix between / being the bootdir
and ROOT being the device name for the root file system;
this allows to deal correctly with root FS on RAM disk.