Do not let gets() warning be fatal to libc.so
Libc includes gets() (both directly and indirectly through
the "ssp" proxy __gets_chk()); this intrinsically unsafe
function is flagged as such at any use; this includes the
building of libc.so.
When static binaries (as was the case so far) or when GCC
is used, things end there.
However,whenusing LLVM/Clang as compiler to produce
dynamically-linked binaries,the flag -Wl,--fatal-warnings
is passed to the linker, and this prevents the building of
libc.so to succeed.
So on MINIX, for the moment, do the same as with GCC and ignore fatality.